+++ indexer/src/java/ru/brandanalyst/indexer/Indexer.java
@@ -72,7 +72,7 @@ public class Indexer implements InitializingBean {
 
         Document doc = new Document();
 
-        doc.add(new Field("id",Integer.toString(b.getId()), Field.Store.YES, Field.Index.NOT_ANALYZED));
+        doc.add(new Field("id",Long.toString(b.getId()), Field.Store.YES, Field.Index.NOT_ANALYZED));
         doc.add(new Field("name",b.getName(), Field.Store.YES, Field.Index.ANALYZED));  // create name
         doc.add(new Field("description",b.getDescription(),Field.Store.YES, Field.Index.ANALYZED));  //create description
         doc.add(new Field("website",b.getWebsite(),Field.Store.YES, Field.Index.ANALYZED));
